% Auto-generated: do not edit by hand
\name{dashGallery}

\alias{dashGallery}

\title{DashGallery component}

\description{

}

\usage{
dashGallery(id=NULL, images=NULL, options=NULL, selected=NULL,
style=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{images}{Unnamed list. Images to display}

\item{options}{Named list. Gallery Options}

\item{selected}{Unnamed list. Selected Images}

\item{style}{Named list. Style}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

